refactor: [Node] подготовлен сервис HandlerService
