+++ node/src/main/java/com/github/jon7even/telegram/BotState.java
@@ -7,20 +7,12 @@ package com.github.jon7even.telegram;
  * @version 2.0
  */
 public enum BotState {
+
     /**
      * Статусы главного меню
      */
     MAIN_START,
     MAIN_HELP,
     MAIN_ITEMS,
-    MAIN_GIFTS,
-
-    /**
-     * Статусы добавления новой компании
-     */
-    COMPANY_NAME,
-    COMPANY_SUM,
-    COMPANY_TYPE_GIFT,
-    COMPANY_IS_GIVEN,
-    COMPANY_IS_DONE;
+    MAIN_GIFTS;
 }
