+++ node/src/main/resources/messages_ru_RU.properties
@@ -1,4 +1,5 @@
-reply.nonSupport=Такая команда еще не поддерживается нашим ботом 😱😱😱
+reply.nonSupport=Такая команда еще не поддерживается нашим ботом. 😱😱😱
+reply.nonSupportedYet=Данная команда еще находится на рефакторинге. 😱
 
 reply.callBackAddNewCompany=Начинаем процесс добавления новой компании. \n\n Давайте начнем с названия компании, введите его:\n\n
